* Changed Grid_Fall (a.k.a. GridFall) to use only bytes in its calculations and as return value.

+++ src/game/game.h
@@ -162,7 +162,7 @@ typedef struct {
 	int (IMPORT *MoveLength) (struct routing_s * map, pos3_t to, qboolean stored);
 	int (IMPORT *MoveNext) (struct routing_s * map, pos3_t from);
 	int (IMPORT *GridHeight) (struct routing_s * map, pos3_t pos);
-	int (IMPORT *GridFall) (struct routing_s * map, pos3_t pos);
+	byte (IMPORT *GridFall) (struct routing_s * map, pos3_t pos);
 	void (IMPORT *GridPosToVec) (struct routing_s * map, pos3_t pos, vec3_t vec);
 	void (IMPORT *GridRecalcRouting) (struct routing_s * map, const char *name, const char **list);
