+++ src/qcommon/cmodel.h
@@ -66,7 +66,7 @@ void Grid_MoveStore(struct routing_s *map);
 int Grid_MoveLength(struct routing_s *map, pos3_t to, qboolean stored);
 int Grid_MoveNext(struct routing_s *map, pos3_t from);
 int Grid_Height(struct routing_s *map, pos3_t pos);
-int Grid_Fall(struct routing_s *map, pos3_t pos);
+byte Grid_Fall(struct routing_s *map, pos3_t pos);
 void Grid_PosToVec(struct routing_s *map, pos3_t pos, vec3_t vec);
